refactor(analytics): move DTOs to src/types/ per core convention
@granit/analytics held its DTOs in feature dirs (widgets/, metrics/) mixed
with runtime type guards, the only domain core package without a src/types/
directory. Relocate all DTO interfaces/types into src/types/ (one file per
domain group + barrel) and keep the runtime type guards in src/widgets/
(guards.ts). The metric-evaluation client stays in src/api/.

Public surface unchanged — src/index.ts re-exports the same names, so
consumers (@granit/react-analytics, react-charts, react-map, dashboards)
and showcase (barrel-only imports) are unaffected.

tsc + ESLint clean; analytics (32) and consumer (213) tests green.
