--- packages/@granit/analytics/src/metrics/index.ts
@@ -1,13 +1,0 @@
-export type {
-  CompareSpec,
-  CompareToken,
-  MetricPreviousPayload,
-  MetricRequest,
-  MetricResponse,
-  MetricSnapshotPayload,
-  PeriodSpec,
-  PeriodToken,
-  RefreshHint,
-  Trend,
-  ValueKind,
-} from './metric-response';
