+++ packages/@granit/analytics/src/types/aggregation.ts
