+++ packages/@granit/analytics/src/types/chart-widget.ts
