+++ packages/@granit/analytics/src/types/kpi-widget.ts
