+++ packages/@granit/analytics/src/types/table-widget.ts
